Finish JobHandler base code (Read desc)
- Finished job handler consequence system
- Added job handler strike system
- Added job handler max tasks system

+++ Town Game/Assets/Scripts/Player/PlayerPropertyHolder.cs
@@ -8,7 +8,7 @@ public class PlayerPropertyHolder : NetworkBehaviour
     [Networked] public NetworkString<_32> nickname { get; set; }
     [Networked] public bool isCultist { get; set; }
     [Networked] public int room { get; set; }
-    [Networked] public int money { get; set; }
+    [Networked] public float money { get; set; }
     [Networked, Capacity(20)] public NetworkLinkedList<int> groups => default;
     [Networked] public int energy { get; set; }
 }
